+++ interactive-run.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash 
 
 GEAR=fw-freeSurfer-clinical-gear
-IMAGE=flywheel/recon-all-clinical:0.2.9
-LOG=recon-all-clinical-0.2.9-668f30a202d8df0cb79ac8f0
+IMAGE=flywheel/recon-all-clinical:0.3.5
+LOG=recon-all-clinical-0.3.5-678115938546c864d8ccc2da
 
 # Command:
 docker run -it --rm --entrypoint bash\
@@ -12,4 +12,5 @@ docker run -it --rm --entrypoint bash\
 	-v /Users/nbourke/GD/atom/unity/fw-gears/${GEAR}/${LOG}/input:/flywheel/v0/input\
 	-v /Users/nbourke/GD/atom/unity/fw-gears/${GEAR}/${LOG}/output:/flywheel/v0/output\
 	-v /Users/nbourke/GD/atom/unity/fw-gears/${GEAR}/${LOG}/work:/flywheel/v0/work\
+	-v /Users/nbourke/GD/atom/unity/fw-gears/${GEAR}/${LOG}/config.json:/flywheel/v0/config.json\
 	$IMAGE